python: moved ecl.ecl.faults to ecl.faults.
python region added, test working.

ecl.faults added fault_segments.

ecl.faults: added all files.

ecl.fault: tests work.

ec.faults: added install check for ecl.grid.
