+++ python/ecl/faults/__init__.py
@@ -1,3 +1,7 @@
+import sys
+if not 'ecl.grid' in sys.modules:
+  raise ImportError("ecl.grid must be imported before ecl.faults.")
+
 from .layer import Layer
 from .fault_collection import FaultCollection
 from .fault import Fault
