+++ python/ecl/faults/fault.py
