+++ python/ecl/faults/fault_segments.py
