+++ python/ecl/faults/layer.py
@@ -16,6 +16,7 @@
 
 import ctypes
 
+from ecl.grid import EclGrid
 from cwrap import BaseCClass
 from ecl.util import monkey_the_camel
 from ecl.util import IntVector
