+++ python/ecl/grid/__init__.py
@@ -5,5 +5,5 @@ import ecl.geo
 
 from .cell import Cell
 from .ecl_grid import EclGrid
-#from .ecl_region import EclRegion
+from .ecl_region import EclRegion
 from .ecl_grid_generator import EclGridGenerator
