+++ python/tests/ecl_tests/test_statoil_faults.py
@@ -20,8 +20,10 @@ except ImportError:
     from unittest import skipIf
 
 import time
-from ecl.ecl.faults import FaultCollection, Fault, FaultLine, FaultSegment
-from ecl.ecl import EclGrid, EclKW, EclDataType
+from ecl import EclDataType
+from ecl.eclfile import EclKW
+from ecl.grid import EclGrid
+from ecl.faults import FaultCollection, Fault, FaultLine, FaultSegment
 from tests import EclTest, statoil_test
